Adds option to hide read items
Implements a new user setting to hide read items from the feed.
This allows users to declutter their view by hiding items that have already been read.

Updates the API to support filtering read items based on the new setting.
The frontend is updated to toggle the setting and reflect the changes in the UI.